Updated wrappers
