+++ src/main/java/me/senseiwells/arucas/api/wrappers/IArucasWrappedClass.java
@@ -1,5 +1,4 @@
 package me.senseiwells.arucas.api.wrappers;
 
 public interface IArucasWrappedClass {
-	String getName();
 }
