+++ src/main/java/me/senseiwells/arucas/utils/Context.java
@@ -3,6 +3,7 @@ package me.senseiwells.arucas.utils;
 import me.senseiwells.arucas.api.ArucasThreadHandler;
 import me.senseiwells.arucas.api.IArucasOutput;
 import me.senseiwells.arucas.api.ISyntax;
+import me.senseiwells.arucas.api.wrappers.ArucasWrapperExtension;
 import me.senseiwells.arucas.api.wrappers.IArucasWrappedClass;
 import me.senseiwells.arucas.throwables.CodeError;
 import me.senseiwells.arucas.throwables.RuntimeError;
@@ -16,7 +17,6 @@ import me.senseiwells.arucas.values.functions.AbstractBuiltInFunction;
 import me.senseiwells.arucas.values.functions.FunctionValue;
 
 import java.util.*;
-import java.util.function.Supplier;
 
 /**
  * Runtime context class of the programming language.
@@ -237,11 +237,16 @@ public class Context {
 	}
 
 	@SuppressWarnings("unused")
-	public ArucasClassValue createWrapperClass(Supplier<IArucasWrappedClass> supplier, List<Value<?>> parameters, ISyntax syntaxPosition) throws CodeError {
-		IArucasWrappedClass wrappedClass = supplier.get();
-		AbstractClassDefinition classDefinition = this.getClassDefinition(wrappedClass.getName());
+	public ArucasClassValue createWrapperClass(Class<? extends IArucasWrappedClass> clazz, List<Value<?>> parameters, ISyntax syntaxPosition) throws CodeError {
+		String wrapperName = ArucasWrapperExtension.getWrapperName(clazz);
+		
+		if (wrapperName == null) {
+			throw new RuntimeError("No such wrapper class exists", syntaxPosition, this);
+		}
+		
+		AbstractClassDefinition classDefinition = this.getClassDefinition(wrapperName);
 		if (classDefinition instanceof WrapperArucasClassDefinition wrappedClassDefinition) {
-			return wrappedClassDefinition.createNewDefinition(wrappedClass, this, parameters, syntaxPosition);
+			return wrappedClassDefinition.createNewDefinition(this, parameters, syntaxPosition);
 		}
 		throw new RuntimeError("No such wrapper class exists", syntaxPosition, this);
 	}
